FE-44: Breadcrumbs done and dusted
