+++ frontend/src/pages/Dashboard/DashboardMain/UploadReady/UploadReady.jsx
@@ -1,33 +1,50 @@
 import axios from 'axios';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../../../../Store/Context';
 import signedDocument from '../../../../assets/images/DashboardImages/upload/signed document.png';
 
 function UploadReady() {
-	const { files, setFiles, getData } = useAuth();
+	const { files, setFiles } = useAuth();
 	console.log(files);
 
-	const url = `https://dev-kcjp.onrender.com/upload`;
+	const url1 = 'http://accountpal.hng.tech:8000/upload_record'
 
-// 	const getData = async () => {
-// 		const formData = new FormData();
-// 		formData.append('file', files);
-// 
-// 		axios
-// 			.post(url, formData, {
-// 				headers: {
-// 					'Content-Type': 'multipart/form-data',
-// 				},
-// 			})
-// 			.then((res) => console.log(res))
-// 			.catch((e) => console.log(e));
-// 	};
+	// const url = `https://dev-kcjp.onrender.com/upload`;
+
+	const getData = async () => {
+			const formData = new FormData();
+			formData.append("file", files);
+
+			axios
+				.post(url1, formData, {
+					headers: {
+						"Content-Type": "multipart/form-data",
+					},
+				})
+				.then((res) => console.log(res?.data))
+				.catch((e) => console.log(e));
+		
+	}
+
+	// 	const getData = async () => {
+	// 		const formData = new FormData();
+	// 		formData.append('file', files);
+	// 
+	// 		axios
+	// 			.post(url, formData, {
+	// 				headers: {
+	// 					'Content-Type': 'multipart/form-data',
+	// 				},
+	// 			})
+	// 			.then((res) => console.log(res))
+	// 			.catch((e) => console.log(e));
+	// 	};
 
 	const navigate = useNavigate();
 	const uploadHandler = () => {
-		
-		getData();
+
+		// getData();
 		navigate('/dashboard/importpage');
 	};
 
@@ -61,7 +78,10 @@ function UploadReady() {
 			</div>
 
 			<button
-				onClick={uploadHandler}
+				onClick={() => {
+					uploadHandler();
+					getData()
+				}}
 				className="bg-[#2E90FA] text-white px-[1.5em] py-[0.8em] rounded-md "
 				type="submit"
 			>
